TreeCovering: working version
The description for Case 3 should be improved:
When we "proceed as in Case 1", as the paper
says, we should in fact finalize each resulting
component. The reason is that the branching
node cannot become a part of another component
simply because then there will be at least two
nodes emanating from a component.